requirements.txt
aioconsole==0.8.1
anyio==3.7.1
bcrypt==4.0.1
beautifulsoup4==4.13.3
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.104.0
Flask==3.1.0
Flask-SocketIO==5.5.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
motor==3.3.1
ogn-client==1.2.1
passlib==1.7.4
ply==3.11
pyasn1==0.6.1
pycryptodomex==3.21.0
pydantic==1.10.13
pymongo==4.11.3
pyserial==3.5
pysnmp-pyasn1==1.1.3
pysnmp-pysmi==1.1.12
pysnmplib==5.0.24
python-dotenv==1.0.0
python-engineio==4.11.2
python-jose==3.3.0
python-multipart==0.0.6
python-socketio==5.12.1
requests==2.32.3
rsa==4.9
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.27.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.23.2
websockets==14.1
Werkzeug==3.1.3
wsproto==1.2.0
